LayoutTests/imported/w3c/web-platform-tests/workers/SharedWorker-extendedLifetime-expected.txt


FAIL SharedWorker lifetime should be extended with extendedLifetime assert_equals: expected (string) "d93750f6-0f5f-47e3-a47d-c1ca5a75dc8c" but got (object) null
